Move some classes to specific package
