+++ src/main/java/net/neoforged/neoforge/client/renderstate/BaseRenderState.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.client;
+package net.neoforged.neoforge.client.renderstate;
 
 import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap;
 import java.util.Map;
